# api.yaml
openapi: 3.0.3
info:
  title: YEMEKİMNEREDE API
  description: |
    YEMEKİMNEREDE uygulaması için RESTful API.

    ## Özellikler
    - Kullanıcı kaydı ve kimlik doğrulama
    - Kullanıcı profil yönetimi
    - Restoran ve menü yönetimi
    - Sipariş oluşturma ve takibi
    - JWT tabanlı kimlik doğrulama
  version: 1.0.0
  contact:
    name: xKRALTR Ekibi
    email: destek@yemekimnerede.com
servers:
  - url: https://api.yemekimnerede.com/v1
    description: Üretim sunucusu
  - url: http://localhost:3000/v1
    description: Geliştirme sunucusu
...
paths:
...
  /cart:
    get:
      tags:
        - orders
      summary: Sepet Görüntüleme
      description: Kullanıcının aktif sepetindeki ürünleri ve toplam tutarı gösterir.
      operationId: getSepet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sepet'
        '401':
          $ref: '#/components/responses/Yetkisiz'

    post:
      tags:
        - orders
      summary: Sepete Ürün Ekleme
      description: Kullanıcının sepetine belirli bir restoranın menüsünden ürün ekler.
      operationId: addSepet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SepetItemInput'
      responses:
        '200':
          description: Ürün sepete eklendi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sepet'
        '400':
          $ref: '#/components/responses/GecersizIstek'
        '401':
          $ref: '#/components/responses/Yetkisiz'
  /cart/{urunId}:
    delete:
      tags:
        - orders
      summary: Sepetten Ürün Çıkarma
      description: Kullanıcının sepetinden belirli bir ürünü kaldırır.
      operationId: removeFromSepet
      security:
        - bearerAuth: []
      parameters:
        - name: urunId
          in: path
          required: true
          description: Sepetten çıkarılacak ürünün kimliği
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ürün sepetten çıkarıldı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sepet'
        '401':
          $ref: '#/components/responses/Yetkisiz'
        '404':
          $ref: '#/components/responses/Bulunamadi'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token ile kimlik doğrulama.
...
  schemas:
...
    SiparisItem:
      type: object
      properties:
        urunId:
          type: string
          format: uuid
        urunAdi:
          type: string
          example: "Adana Kebap"
        adet:
          type: integer
          example: 2
        birimFiyat:
          type: number
          format: float
          example: 180.00
        toplamFiyat:
          type: number
          format: float
          example: 360.00
...
    Sepet:
      type: object
      properties:
        kullaniciId:
          type: string
          format: uuid
          example: "usr-123"
        restoranId:
          type: string
          format: uuid
          nullable: true
          example: "rst-456"
        urunler:
          type: array
          items:
            $ref: '#/components/schemas/SiparisItem'
        araToplam:
          type: number
          format: float
          example: 360.00
    SepetItemInput:
      type: object
      required:
        - restoranId
        - urunId
        - adet
      properties:
        restoranId:
          type: string
          format: uuid
          example: "rst-456"
        urunId:
          type: string
          format: uuid
          example: "urun-456"
        adet:
          type: integer
          minimum: 1
          example: 2
...
    Hata:
      type: object
      required:
        - kod
        - mesaj
      properties:
        kod:
          type: string
          description: Hata kodu
          example: "GECERSIZ_ISTEK"
        mesaj:
          type: string
          description: Hata mesajı
          example: "Geçersiz e-posta formatı"
        detaylar:
          type: array
          items:
            type: object
            properties:
              alan:
                type: string
                example: "email"
              mesaj:
                type: string
                example: "E-posta formatı geçersiz"
  responses:
    GecersizIstek:
      description: Geçersiz istek
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Hata'
          example:
            kod: "GECERSIZ_ISTEK"
            mesaj: "İstek parametreleri geçersiz"
    Yetkisiz:
      description: Yetkisiz erişim - geçerli bir JWT token gereklidir
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Hata'
          example:
            kod: "YETKISIZ"
            mesaj: "Kimlik doğrulama başarısız"
    Bulunamadi:
      description: İstenen kaynak bulunamadı
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Hata'
          example:
            kod: "BULUNAMADI"
            mesaj: "İstenen kaynak bulunamadı"